Cleanup StagingBuffer usage in VkEmulation
Staging buffer in VkEmulation is used for reading back
and updating color buffer data, and is not actually
allocated as an external memory. CL makes buffer creation
and usage cleaner with a self contained structure, which
can later be used for other staging buffers, or for
implementing multiple staging buffers when needed.

Bug: 389637419
Test: run XR emulator on multiple platforms
Change-Id: I9eb0bdea496539407abb908360414f72c28acba6