Handle timeline semaphores on virtual queue
Special handling of queue submissions where the signalling will be done
after a wait operation sent to the queue.

When a single physical queue is shared with VulkanVirtualQueue feature,
wait operations can block the physical queue and the subsequent signals
cannot be processed. We defer the real dispatch of such submissions until
another queue submission with the necessary semaphore signaling is made.
For other post-submit operations, we treat the pending ops as if they
have been submitted, because the object lifetime management (e.g. fences
semaphores, command buffers) should be controlled by the app for valid
usage and the fence for the host operations won't be signalled until the
final execution chain is completed.

The current implementation does not support presentation images and the
cope operations to GL will still wait on the fence immediately.

Bug: 398986781
Test: dEQP-VK.synchronization2.timeline_semaphore.wait_before_signal.*
Change-Id: Ifedfd1ff690e5c6c4263dce8374533d24fe1bc14