Re-implement Gfxstream pipes in new VirtioGpuPipe.[cpp|h]
... which will allow us to delete the pipe implemententations in
host/gl/gl-host-common and remove another dependency on
hardward/google/aemu/host-common.

The use of the `std::shared_ptr<VirtioGpuPipe>` also lets us remove some
of the awkwardness of the `resetPipe()` which was needed to re-associate
pipes with all resources in a context after the pipe was converted from
the "initial pipe" to an actual pipe implementation.

NO_IFTTT=host pipe snapshot/restore is a pre-existing TODO

Bug: b/403363880
Test: cvd create --gpu_mode=gfxstream
Test: cvd create --gpu_mode=gfxstream_guest_angle
Change-Id: I27b60c513317f26c2eec0ec7970cbe117f38575c