[vulkan snapshot]: remove duplicate api traces
Currently, when an api creates multiple handles
the api trace is saved and loaded multiple times.

It happended to work, likely due to the fact most
apis happen to create just 1 handle.
but is wrong. And it leads to crashes due to handle
messing up during load.
e.g., when creating descriptors, some could create
128 in one call.

Bug: 418814845

Change-Id: Ie5c153789dc62deb89414cf34adaa5355773181c