Restore some default behavior from before ag/32981102
ag/32981102 was performing some needed refactoring to decouple
gfxstream from AEMU in order to support standalone builds of
gfxstream for Cuttlefish's move from AOSP to github. However,
the change accidentally removed some default behavior which
controls how contexts are used.

This was not caught earlier because most of CI testing uses
Nvidia GPUs which seem to be more permissive...

Bug: b/413575758
Test: cvd create --gpu_mode=gfxstream w/ AMD GPUs
Change-Id: Ie44a921e1d2745e6402c7117aeed5088ef52414b