[vulkan snapshot]: simplify vk update descriptor
Currently, we save all the vkUpdateDescriptorSets commands
and replay then during loading.

This leads to very large amount of commands and it is
slowing down both the saving and loading, when the emulator
has run a longer period of time.

Prior work worked around this problem by saving only the
active descriptors, but it only works with VulkanBatchedDescriptorSetUpdate
feature.

This cl improves prior work to make it work without
VulkanBatchedDescriptorSetUpdate feature.

Bug: 417788798

Change-Id: I5b83566db3723135b9d8dd2df56ad1d7f2a751ed