Enable running emulator without emulationGl
Do not initialize emulationGl when vulkan composition is used and
handle usages of it by reporting errors to the logs with the function
name when possible, rather than doing a fatal crash.
Calls to createEmulatedEglFenceSync are still being made frequently
in GuestAngle mode, probably due to a separate bug, so the error
message is logged only once.
Also changes the return type of EmulatedEglImage::destroy to a regular
bool to limit usages of EGL structures.

Bug: 389646068
Test: -feature GuestAngle,VulkanNativeSwapchain,GuestUsesAngle
Change-Id: Ib9f7f1b98fe91aa2fcd47bd98f4494eb947d00b7