Wait for the thread completion in the thread destructor (win32).
This is also the behavior of the thread pthread version.

This issue is found in the android emulator sources. However, this
copy from gfxstream is exactly the same code. Below is a description
of the issue found in the android emulator.

When running android-emu_unittests on Windows, sometime there is a
crash without any crash dumps. The exit code is either 0xC0000005
or 0xC0000409. When happened, it was always running adb related
unit tests.
The crash trigger is the following code behavior:
TEST(adbXXXXX, testYYYY) {
    TestConnectThread testThread(testListener.port(), 2);
    ...................
}
At the end of the test function, destructor of android::base::thread
is called. Note that the thread could still be running. Here is the
destructor:
Thread::~Thread() {
    if (mThread) {
        assert(!mStarted || mFinished);
        CloseHandle(mThread);
    }
}
Adding to Windows API description, CloseHandle does NOT wait for NOR
kill the target thread. It merely closes the handle. This destructor
closes the handle and frees the thread object.
However, the thread object is still referenced by the thread:
DWORD WINAPI Thread::thread_main(void* arg) {
    {
        // no need to call maskAllSignals() here: we know
        // that on Windows it's a noop
        Thread* self = reinterpret_cast<Thread*>(arg);
        auto ret = self->main();

        {
            AutoLock lock(self->mLock);
            self->mFinished = true;
            self->mExitStatus = ret;
        }

        self->onExit();
        // |self| is not valid beyond this point
    }

    // Ensure all thread-local values are released for this thread.
    ::android::base::ThreadStoreBase::OnThreadExit();

    // This return value is ignored.
    return 0;
}
After the destructor, the "self" pointer becomes a dangling pointer. What's more,
self->onExit() could call arbitrary code.

Change-Id: If26f210df6efe9d87024cb52b6c22fc249471cc7