Implement HostAllocation external memory mode
Add new external memory mode and enable it for windows to be able
to use host allocations and relevant import operations to provide
external memory support.

This removes dependency on win32 mode support to enable vulkan
composition with lavapipe and swiftshader on windows. Note that
enablement for ahb vulkan extension on the guest requires a system
side update to properly check the external memory support.

Bug: 389646068
Bug: 409769371
Test: -gpu lavapipe -feature VulkanNativeSwapchain on windows
Change-Id: I5105550125ed3f4eb31c566a090cf3b1dcaf1410